feat: introduce generic handler, use at settlement endpoints (#25)
* feat!: create generic endpoint handler

* feat: refactor settlement endpoints until update one

* refactor: all settlement endpoints use generic handler

+++ src/ExpenseSplitter.Api.Application/Settlements/DeleteSettlement/DeleteSettlementCommandHandler.cs
@@ -1,39 +1,36 @@
 using ExpenseSplitter.Api.Application.Abstractions.Authentication;
 using ExpenseSplitter.Api.Application.Abstractions.Cqrs;
+using ExpenseSplitter.Api.Application.Abstractions.Etag;
 using ExpenseSplitter.Api.Domain.Abstractions;
 using ExpenseSplitter.Api.Domain.Settlements;
 
 namespace ExpenseSplitter.Api.Application.Settlements.DeleteSettlement;
 
-public class DeleteSettlementCommandHandler : ICommandHandler<DeleteSettlementCommand>
+public class DeleteSettlementCommandHandler(
+    ISettlementRepository settlementRepository,
+    IUserContext userContext,
+    IUnitOfWork unitOfWork,
+    IEtagService etagService
+) : ICommandHandler<DeleteSettlementCommand>
 {
-    private readonly ISettlementRepository settlementRepository;
-    private readonly IUserContext userContext;
-    private readonly IUnitOfWork unitOfWork;
-
-    public DeleteSettlementCommandHandler(
-        ISettlementRepository settlementRepository,
-        IUserContext userContext,
-        IUnitOfWork unitOfWork
-    )
-    {
-        this.settlementRepository = settlementRepository;
-        this.userContext = userContext;
-        this.unitOfWork = unitOfWork;
-    }
-
     public async Task<Result> Handle(DeleteSettlementCommand request, CancellationToken cancellationToken)
     {
         var settlementId = new SettlementId(request.Id);
         var settlement = await settlementRepository.GetById(settlementId, cancellationToken);
+
         if (settlement is null)
         {
-            return Result.Failure<Guid>(SettlementErrors.NotFound);
+            return Result.Failure(SettlementErrors.NotFound);
+        }
+        
+        if (etagService.HasIfMatchConflict(settlement))
+        {
+            return Result.Failure(SettlementErrors.IfMatchHeaderConflict);
         }
 
         if (settlement.CreatorUserId != userContext.UserId)
         {
-            return Result.Failure<Guid>(SettlementErrors.Forbidden);
+            return Result.Failure(SettlementErrors.Forbidden);
         }
 
         settlementRepository.Remove(settlement);
